Fixes for debug output, new pixel buffer
Added:
PixelBuffer class. This stores a texture and a chunk of memory that stores the values of each pixel. This allows for easy drawing of pixels on the screen.

Fixed:
Added -mwindows and -Wl,-subsystem,console flags to linker in cmake, so now there's visible console output for easier debugging.

Changes:
Changed the 'demo' from a moving square to a simple 'painting' program using the pixel buffer. Click to draw pixels.
Added ImGuiIO as an argument to HandleEvents so you can check, for example, if the mouse is above an imgui window.